+++ features/user/topchat/src/main/java/com/tokopedia/topchat/chatroom/view/viewmodel/WebsocketAttachmentContract.kt
@@ -14,7 +14,7 @@ open class WebsocketAttachmentContract (
 
 @Keep
 open class WebsocketAttachmentData (
-        val message_id: Int,
+        val message_id: Long,
         val message: String,
         val source: String,
         val attachment_type: Int,
